+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 2.3.6 (2018-09-28)
+
+    implementation 'com.github.uizaio:uiza-android-sdk-player:2.3.6'
+- Add sample custom skin via using code.
+- Add more API for UZVideo.
 
 ### 2.3.5 (2018-09-27)